app: move GET /api/ota to app.routers.ota
